+++ _apps/default/view/503.php
@@ -1,0 +1,9 @@
+<?php
+if (!defined('ROOT')) {
+    require_once $_SERVER['ROOT_PATH'] . $_SERVER['ERROR_PATH'];
+}
+?>
+<div class="main">
+    <h1>503 Error</h1>
+    <h1>Under Construction Or Down For Maintenance</h1>
+</div>